[Style] Observed / Model 분리
